+++ examples/pytorch/model_zoo/chem/generative_models/jtnn/jtnn/__init__.py
@@ -1,6 +1,4 @@
 from .mol_tree import Vocab
-from .jtnn_vae import DGLJTNNVAE
-from .mpn import DGLMPN
 from .nnutils import cuda
 from .datautils import JTNNDataset, JTNNCollator
 from .chemutils import decode_stereo
