+++ python/dgl/model_zoo/chem/jtnn/jtnn/__init__.py
@@ -2,5 +2,4 @@ from .mol_tree import Vocab
 from .jtnn_vae import DGLJTNNVAE
 from .mpn import DGLMPN
 from .nnutils import create_var, cuda
-from .datautils import JTNNDataset, JTNNCollator
 from .chemutils import decode_stereo
